webkit/data/layout_tests/chrome/fast/dom/xss-DENIED-javascript-variations-expected.txt

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL data:text/html,<p>Inner text should not be replaced.<p> from frame with URL xss-DENIED-javascript-variations.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL data:text/html,<p>Inner text should not be replaced.<p> from frame with URL xss-DENIED-javascript-variations.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL data:text/html,<p>Inner text should not be replaced.<p> from frame with URL xss-DENIED-javascript-variations.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL data:text/html,<p>Inner text should not be replaced.<p> from frame with URL xss-DENIED-javascript-variations.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL data:text/html,<p>Inner text should not be replaced.<p> from frame with URL xss-DENIED-javascript-variations.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL data:text/html,<p>Inner text should not be replaced.<p> from frame with URL xss-DENIED-javascript-variations.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL data:text/html,<p>Inner text should not be replaced.<p> from frame with URL xss-DENIED-javascript-variations.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL data:text/html,<p>Inner text should not be replaced.<p> from frame with URL xss-DENIED-javascript-variations.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL data:text/html,<p>Inner text should not be replaced.<p> from frame with URL xss-DENIED-javascript-variations.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL data:text/html,<p>Inner text should not be replaced.<p> from frame with URL xss-DENIED-javascript-variations.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL data:text/html,<p>Inner text should not be replaced.<p> from frame with URL xss-DENIED-javascript-variations.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL data:text/html,<p>Inner text should not be replaced.<p> from frame with URL xss-DENIED-javascript-variations.html. Domains, protocols and ports must match.




--------
Frame: 'DUPA'
--------
Inner text should not be replaced.
